src/retrieval/requirements.txt
langchain==0.1.11
transformers
sentence_transformers==2.2.2
langchain-community==0.0.27
langchain_experimental==0.0.53
InstructorEmbedding
faiss-cpu
google-cloud-storage